refactor: extract IDEWindowManager from SikulixIDE (Phase 3, Step 3)
Move toolbar buttons (ButtonInsertImage, ButtonSubregion, ButtonLocation,
ButtonOffset, ButtonShow, ButtonShowIn, ButtonRecord), message area
(initMessageArea, clearMessageArea, collapse/uncollapse, toggleCollapsed),
tabs (initTabs, initShortcutKeys), capture hotkeys, toolbar init, and
layout building (buildLayout from startGUI) to IDEWindowManager.java.

Cleaned 12 orphan imports from SikulixIDE. Added missing
SikulixFileChooser import to IDEWindowManager.

SikulixIDE.java: 2196 -> 1648 lines (-548)
IDEWindowManager.java: 667 lines (new)
